Move elastic task into existing kf-pytorch plugin
