--- plugins/flytekit-torch-elastic/flytekitplugins/torchelastic/__init__.py
@@ -1,13 +1,0 @@
-"""
-.. currentmodule:: flytekitplugins.torchelastic
-
-This package contains things that are useful when extending Flytekit.
-
-.. autosummary::
-   :template: custom.rst
-   :toctree: generated/
-
-   Torch-Elastic
-"""
-
-from .task import Elastic
