--- plugins/flytekit-torch-elastic/setup.py
@@ -1,36 +1,0 @@
-from setuptools import setup
-
-PLUGIN_NAME = "torchelastic"
-
-microlib_name = f"flytekitplugins-{PLUGIN_NAME}"
-
-plugin_requires = ["flytekit>=1.3.0b2,<2.0.0", "flyteidl>=0.21.4"]
-
-__version__ = "0.0.0+develop"
-
-setup(
-    name=microlib_name,
-    version=__version__,
-    author="flyteorg",
-    author_email="admin@flyte.org",
-    description="K8s based torch-elastic plugin for flytekit",
-    namespace_packages=["flytekitplugins"],
-    packages=[f"flytekitplugins.{PLUGIN_NAME}"],
-    install_requires=plugin_requires,
-    license="apache2",
-    python_requires=">=3.6",
-    classifiers=[
-        "Intended Audience :: Science/Research",
-        "Intended Audience :: Developers",
-        "License :: OSI Approved :: Apache Software License",
-        "Programming Language :: Python :: 3.7",
-        "Programming Language :: Python :: 3.8",
-        "Programming Language :: Python :: 3.9",
-        "Programming Language :: Python :: 3.10",
-        "Topic :: Scientific/Engineering",
-        "Topic :: Scientific/Engineering :: Artificial Intelligence",
-        "Topic :: Software Development",
-        "Topic :: Software Development :: Libraries",
-        "Topic :: Software Development :: Libraries :: Python Modules",
-    ],
-)
